EduConnect_Lesotho_Dig_Library/requirements.txt
# Core Flask dependencies
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-WTF==1.1.1
Flask-Mail==0.9.1

# Database drivers
PyMySQL==1.1.0
# mysqlclient==2.2.0  # Commented out - can cause issues on Windows, PyMySQL is sufficient

# Security and forms
WTForms==3.0.1
Werkzeug==2.3.7
cryptography==41.0.4

# File handling and utilities
Pillow==11.0.0
python-dotenv==1.0.0

# Date and time handling
python-dateutil==2.8.2

# Development and testing
pytest==7.4.2
pytest-flask==1.2.0
flask-testing==0.8.1
requests==2.31.0

# Production server
gunicorn==21.2.0

# Optional: For enhanced features
redis==4.6.0  # For caching (optional)
celery==5.3.1  # For background tasks (optional)